requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
boto3
requests
pydantic
strands
strands-agent